pyproject.toml
[tool.poetry]
name = "eindfeest"
version = "0.1.0"
description = ""
authors = ["M-spree <135598804+M-spree@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "eindfeest", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.26.2"
matplotlib = "^3.8.2"
pyvisa-py = "^0.7.1"
pyserial = "^3.5"
pyside6 = "^6.6.1"
pyqtgraph = "^0.13.3"


[tool.poetry.group.dev.dependencies]
nsp2visasim = "^1.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
diode = "eindfeest.diode-experiment:scan"